separete tests for different locales #996
